requirements.txt
customtkinter
pillow
rembg
opencv-python
numpy
matplotlib
